auth-center/python/requirements.txt
flask
python-dotenv
qrcode[pil]
requests
PyNaCl
base58
